+++ tbx/layout/+uix/BoxPanel.m
@@ -10,7 +10,7 @@ classdef BoxPanel < uix.Panel
     %
     %  See also: uix.Panel, uipanel, uix.CardPanel
 
-    %  Copyright 2009-2024 The MathWorks, Inc.
+    %  Copyright 2009-2025 The MathWorks, Inc.
 
     properties( Access = public, Dependent, AbortSet )
         TitleColor % title background color [RGB]
@@ -24,16 +24,10 @@ classdef BoxPanel < uix.Panel
         CloseRequestFcn % close request callback
     end
 
-    properties( Dependent, SetAccess = private )
-        TitleHeight % title panel height [pixels]
-    end
-
     properties( Access = private )
         TitleBar % title bar
         TitleText % title text
-        Title_ = '' % cache of title
         TitleAccess = 'public' % 'private' when getting or setting Title, 'public' otherwise
-        TitleHeight_ = -1 % cache of title text height (-1 denotes stale cache)
         MinimizeButton % button
         MaximizeButton % button
         DockButton % button
@@ -47,8 +41,6 @@ classdef BoxPanel < uix.Panel
 
     properties( Constant, Access = private )
         NullTitle = char.empty( [2 0] ) % an obscure empty string, the actual panel Title
-        BlankTitle = ' ' % a non-empty blank string, the empty uicontrol String
-        DummyControl = matlab.ui.control.UIControl() % dummy uicontrol
     end
 
     properties( Access = public, Dependent, AbortSet )
@@ -69,9 +61,15 @@ classdef BoxPanel < uix.Panel
         CloseTooltipString % now CloseTooltip
     end % deprecated
 
+    properties( Access = public, Dependent, Hidden )
+        TitleColor_I % backing for TitleColor
+        TextColor % ForegroundColor companion
+        TextColor_I % backing for TextColor
+        TextColorMode % TextColor mode [auto|manual]
+    end
+
     properties( Access = public, Hidden )
-        TitleColor_I = [0 0.251 0.451] % backing for TitleColor
-        TitleColorMode = 'auto' % TitleColor mode [auto|manual]
+        TitleColorMode % TitleColor mode [auto|manual]
     end
 
     events( Hidden, NotifyAccess = private )
@@ -94,60 +92,57 @@ classdef BoxPanel < uix.Panel
             %  v1, etc.
 
             % Define default colors
-            foregroundColor = get( 0, 'DefaultUipanelForegroundColor' );
-            backgroundColor = obj.TitleColor_I;
+            foregroundColor = [0.1294 0.1294 0.1294]; % --mw-color-primary
+            titleColor = [0.8706 0.9373 1.0000]; % --mw-backgroundColor-chatBubble
 
             % Create title bar
             titleBar = uix.HBox( 'Internal', true, 'Parent', obj, ...
-                'Units', 'pixels', 'BackgroundColor', backgroundColor );
-
-            % Create title text
+                'Units', 'pixels', 'BackgroundColor', titleColor );
             titleText = uicontrol( 'Parent', titleBar, ...
-                'Style', 'text', 'String', obj.BlankTitle, ...
-                'HorizontalAlignment', 'left', ...
+                'Style', 'text', 'HorizontalAlignment', 'left', ...
                 'ForegroundColor', foregroundColor, ...
-                'BackgroundColor', backgroundColor );
+                'BackgroundColor', titleColor ); % for Java
 
             % Create buttons
             minimizeButton = uicontrol( 'Parent', [], ...
                 'Style', 'text', 'HorizontalAlignment', 'center', ...
                 'ForegroundColor', foregroundColor, ...
-                'BackgroundColor', backgroundColor, ...
+                'BackgroundColor', titleColor, ...
                 'Enable', 'on', 'FontWeight', 'bold', ...
                 'String', char( 9652 ), ...
                 'TooltipString', 'Collapse this panel' );
             maximizeButton = uicontrol( 'Parent', [], ...
                 'Style', 'text', 'HorizontalAlignment', 'center', ...
                 'ForegroundColor', foregroundColor, ...
-                'BackgroundColor', backgroundColor, ...
+                'BackgroundColor', titleColor, ...
                 'Enable', 'on', 'FontWeight', 'bold', ...
                 'String', char( 9662 ), ...
                 'TooltipString', 'Expand this panel' );
             dockButton = uicontrol( 'Parent', [], ...
                 'Style', 'text', 'HorizontalAlignment', 'center', ...
                 'ForegroundColor', foregroundColor, ...
-                'BackgroundColor', backgroundColor, ...
+                'BackgroundColor', titleColor, ...
                 'Enable', 'on', 'FontWeight', 'bold', ...
                 'String', char( 8600 ), ...
                 'TooltipString', 'Dock this panel' );
             undockButton = uicontrol( 'Parent', [], ...
                 'Style', 'text', 'HorizontalAlignment', 'center', ...
                 'ForegroundColor', foregroundColor, ...
-                'BackgroundColor', backgroundColor, ...
+                'BackgroundColor', titleColor, ...
                 'Enable', 'on', 'FontWeight', 'bold', ...
                 'String', char( 8599 ), ...
                 'TooltipString', 'Undock this panel' );
             helpButton = uicontrol( 'Parent', [], ...
                 'Style', 'text', 'HorizontalAlignment', 'center', ...
                 'ForegroundColor', foregroundColor, ...
-                'BackgroundColor', backgroundColor, ...
+                'BackgroundColor', titleColor, ...
                 'Enable', 'on', 'FontWeight', 'bold', ...
                 'String', '?', ...
                 'TooltipString', 'Get help on this panel' );
             closeButton = uicontrol( 'Parent', [], ...
                 'Style', 'text', 'HorizontalAlignment', 'center', ...
                 'ForegroundColor', foregroundColor, ...
-                'BackgroundColor', backgroundColor, ...
+                'BackgroundColor', titleColor, ...
                 'Enable', 'on', 'FontWeight', 'bold', ...
                 'String', char( 215 ), ...
                 'TooltipString', 'Close this panel' );
@@ -162,6 +157,9 @@ classdef BoxPanel < uix.Panel
             obj.UndockButton = undockButton;
             obj.HelpButton = helpButton;
             obj.CloseButton = closeButton;
+
+            % Synchronize ForegroundColor without flipping mode
+            obj.ForegroundColor_I = foregroundColor;
 
             % Create listeners
             addlistener( obj, 'BorderWidth', 'PostSet', ...
@@ -231,18 +229,23 @@ classdef BoxPanel < uix.Panel
 
         end % set.TitleColor
 
+        function value = get.TitleColor_I( obj )
+
+            value = obj.TitleBar.BackgroundColor;
+
+        end % get.TitleColor_I
+
         function set.TitleColor_I( obj, value )
 
-            try
-                obj.DummyControl.ForegroundColor = value; % colorspec
-                value = obj.DummyControl.ForegroundColor; % rgb
-                obj.TitleColor_I = value; % store
-                obj.redrawTitle() % apply
-            catch
-                throwAsCaller( MException( 'uix:InvalidPropertyValue', ...
-                    'Property ''TitleColor_I'' must be a colorspec.' ) )
-            end
-            obj.TitleText.ForegroundColor = obj.ForegroundColor;
+            % Set underlying
+            obj.TitleBar.BackgroundColor = value;
+            obj.TitleText.BackgroundColor = value;
+            obj.MinimizeButton.BackgroundColor = value;
+            obj.MaximizeButton.BackgroundColor = value;
+            obj.DockButton.BackgroundColor = value;
+            obj.UndockButton.BackgroundColor = value;
+            obj.HelpButton.BackgroundColor = value;
+            obj.CloseButton.BackgroundColor = value;
 
         end % set.TitleColor_I
 
@@ -258,6 +261,54 @@ classdef BoxPanel < uix.Panel
             end
 
         end % set.TitleColorMode
+
+        function value = get.TextColor( obj )
+
+            value = obj.TextColor_I; % delegate
+
+        end % get.TextColor
+
+        function set.TextColor( obj, value )
+
+            obj.TextColor_I = value; % apply
+            obj.TextColorMode = 'manual'; % flip mode
+            obj.ForegroundColorMode = 'manual'; % mirror mode
+
+        end % set.TextColor
+
+        function value = get.TextColor_I( obj )
+
+            value = obj.TitleText.ForegroundColor; % get underlying
+
+        end % get.TextColor_I
+
+        function set.TextColor_I( obj, value )
+
+            % Set underlying
+            obj.TitleText.ForegroundColor = value;
+            obj.MinimizeButton.ForegroundColor = value;
+            obj.MaximizeButton.ForegroundColor = value;
+            obj.DockButton.ForegroundColor = value;
+            obj.UndockButton.ForegroundColor = value;
+            obj.HelpButton.ForegroundColor = value;
+            obj.CloseButton.ForegroundColor = value;
+
+            % Mirror
+            obj.ForegroundColor_I = value;
+
+        end % set.TextColor_I
+
+        function value = get.TextColorMode( obj )
+
+            value = obj.ForegroundColorMode; % delegate
+
+        end % get.TextColorMode
+
+        function set.TextColorMode( obj, value )
+
+            obj.ForegroundColorMode = value; % delegate
+
+        end % set.TextColorMode
 
         function value = get.Minimized( obj )
 
@@ -377,7 +428,6 @@ classdef BoxPanel < uix.Panel
                     'DockFcn' ) )
             end
 
-
             % Update buttons
             obj.redrawButtons()
 
@@ -449,12 +499,6 @@ classdef BoxPanel < uix.Panel
             obj.redrawButtons()
 
         end % set.CloseRequestFcn
-
-        function value = get.TitleHeight( obj )
-
-            value = obj.TitleBar.Position(4);
-
-        end % get.TitleHeight
 
     end % accessors
 
@@ -699,6 +743,7 @@ classdef BoxPanel < uix.Panel
         function onFontAngleChanged( obj, ~, ~ )
             %onFontAngleChanged  Event handler for FontAngle changes
 
+            % Set
             obj.TitleText.FontAngle = obj.FontAngle;
 
         end % onFontAngleChanged
@@ -708,10 +753,6 @@ classdef BoxPanel < uix.Panel
 
             % Set
             obj.TitleText.FontName = obj.FontName;
-
-            % Mark as dirty
-            obj.TitleHeight_ = -1;
-            obj.Dirty = true;
 
         end % onFontNameChanged
 
@@ -729,7 +770,6 @@ classdef BoxPanel < uix.Panel
             obj.CloseButton.FontSize = fontSize;
 
             % Mark as dirty
-            obj.TitleHeight_ = -1;
             obj.Dirty = true;
 
         end % onFontSizeChanged
@@ -737,6 +777,7 @@ classdef BoxPanel < uix.Panel
         function onFontUnitsChanged( obj, ~, ~ )
             %onFontUnitsChanged  Event handler for FontUnits changes
 
+            % Set
             fontUnits = obj.FontUnits;
             obj.TitleText.FontUnits = fontUnits;
             obj.HelpButton.FontUnits = fontUnits;
@@ -749,6 +790,7 @@ classdef BoxPanel < uix.Panel
         function onFontWeightChanged( obj, ~, ~ )
             %onFontWeightChanged  Event handler for FontWeight changes
 
+            % Set
             obj.TitleText.FontWeight = obj.FontWeight;
 
         end % onFontWeightChanged
@@ -756,14 +798,7 @@ classdef BoxPanel < uix.Panel
         function onForegroundColorChanged( obj, ~, ~ )
             %onForegroundColorChanged  Event handler for ForegroundColor changes
 
-            foregroundColor = obj.ForegroundColor;
-            obj.TitleText.ForegroundColor = foregroundColor;
-            obj.MinimizeButton.ForegroundColor = foregroundColor;
-            obj.MaximizeButton.ForegroundColor = foregroundColor;
-            obj.DockButton.ForegroundColor = foregroundColor;
-            obj.UndockButton.ForegroundColor = foregroundColor;
-            obj.HelpButton.ForegroundColor = foregroundColor;
-            obj.CloseButton.ForegroundColor = foregroundColor;
+            obj.TextColor = obj.ForegroundColor;
 
         end % onForegroundColorChanged
 
@@ -772,7 +807,7 @@ classdef BoxPanel < uix.Panel
 
             if strcmp( obj.TitleAccess, 'public' ) && isjsdrawing() == false
                 obj.TitleAccess = 'private'; % start
-                obj.Title = obj.Title_;
+                obj.Title = obj.TitleText.String; % transfer
             end
 
         end % onTitleReturning
@@ -781,7 +816,7 @@ classdef BoxPanel < uix.Panel
             %onTitleReturned  Event handler for Title changes
 
             if isjsdrawing() == false
-                obj.Title = obj.NullTitle; % unset Title
+                obj.Title = obj.NullTitle; % unset
             end
             obj.TitleAccess = 'public'; % finish
 
@@ -794,19 +829,9 @@ classdef BoxPanel < uix.Panel
 
                 % Set
                 obj.TitleAccess = 'private'; % start
-                title = obj.Title;
-                obj.Title_ = title;
-                if isempty( title )
-                    obj.TitleText.String = obj.BlankTitle;
-                else
-                    obj.TitleText.String = [' ' title]; % set String to title
-                end
-                obj.Title = obj.NullTitle; % unset Title
+                obj.TitleText.String = obj.Title; % transfer
+                obj.Title = obj.NullTitle; % unset
                 obj.TitleAccess = 'public'; % finish
-
-                % Mark as dirty
-                obj.TitleHeight_ = -1;
-                obj.Dirty = true;
 
             end
 
@@ -887,32 +912,38 @@ classdef BoxPanel < uix.Panel
             %
             %  p.redraw() redraws the panel.
 
-            % Compute positions
-            bounds = hgconvertunits( ancestor( obj, 'figure' ), ...
+            % Compute available space
+            iB = hgconvertunits( ancestor( obj, 'figure' ), ...
                 [0 0 1 1], 'normalized', 'pixels', obj );
-            tX = 1;
-            tW = max( bounds(3), 1 );
-            tH = obj.TitleHeight_; % title height
-            if tH == -1 % cache stale, refresh
-                tE = uix.extent( obj.TitleText ); % extent
-                tH = tE(4); % required height
-                obj.TitleHeight_ = tH; % store
-                obj.TitleBar.Widths(2:end) = tH; % square buttons
-            end
-            tY = 1 + bounds(4) - tH;
-            p = obj.Padding_;
-            cX = 1 + p;
-            cW = max( bounds(3) - 2 * p, 1 );
-            cH = max( bounds(4) - tH - 2 * p, 1 );
-            cY = tY - p - cH;
-            contentsPosition = [cX cY cW cH];
+
+            % Compute title bar position
+            tX = 1; % full width
+            tW = iB(3); % full width
+            [tA, tD] = fontmetrics( obj ); % ascent and descent
+            tP = 0.1 * ( tD - tA ); % padding
+            tP = min( tP, tA ); % not more than ascent
+            tH = tD + tP;
+            tY = 1 + iB(4) - tD + tA - 2 * tP;
+
+            % Compute contents position
+            cP = obj.Padding_;
+            cX = 1 + cP;
+            cW = iB(3) - 2 * cP;
+            cW = max( cW, 0 ); % nonnegative
+            cH = iB(4) - tD + tA - 2 * tP - 2 * cP;
+            cH = max( cH, 0 ); % nonnegative
+            cY = 1 + iB(4) - tD + tA - 2 * tP - cH - cP;
+
+            % Redraw title bar
+            titleBar = obj.TitleBar;
+            uix.setPosition( titleBar, [tX tY tW tH], 'pixels' )
+            titleBar.Widths(2:end) = tH; % square buttons
 
             % Redraw contents
             contents = obj.Contents_;
             for ii = 1:numel( contents )
-                uix.setPosition( contents(ii), contentsPosition, 'pixels' )
+                uix.setPosition( contents(ii), [cX cY cW cH], 'pixels' )
             end
-            obj.TitleBar.Position = [tX tY tW tH];
 
         end % redraw
 
@@ -946,6 +977,7 @@ classdef BoxPanel < uix.Panel
 
             map = getThemeMap@uix.Panel();
             map.ForegroundColor = '--mw-color-primary';
+            map.TextColor = '--mw-color-primary';
             map.TitleColor = '--mw-backgroundColor-chatBubble';
 
         end % getThemeMap
@@ -1001,26 +1033,9 @@ classdef BoxPanel < uix.Panel
             end
 
             % Set sizes
-            if obj.TitleHeight_ > 0
-                obj.TitleBar.Widths(2:end) = obj.TitleHeight_;
-            end
+            obj.TitleBar.Widths(2:end) = obj.TitleBar.Position(4);
 
         end % redrawButtons
-
-        function redrawTitle( obj )
-            %redrawTitle  Redraw title bar
-
-            color = obj.TitleColor_I;
-            obj.TitleBar.BackgroundColor = color;
-            obj.TitleText.BackgroundColor = color;
-            obj.MinimizeButton.BackgroundColor = color;
-            obj.MaximizeButton.BackgroundColor = color;
-            obj.DockButton.BackgroundColor = color;
-            obj.UndockButton.BackgroundColor = color;
-            obj.HelpButton.BackgroundColor = color;
-            obj.CloseButton.BackgroundColor = color;
-
-        end % redrawTitle
 
     end % helper methods
 
@@ -1037,3 +1052,113 @@ for ii = 1:numel( s )
 end
 
 end % isjsdrawing
+
+function [a, d] = fontmetrics( obj )
+%fontmetrics  Font metrics
+%
+%  [a,d] = fontmetrics(b) returns the distances from the top to the ascent
+%  a and the descent d in pixels.
+
+% Java font metrics
+ja = [ ...
+       0     0     1     3     4     5     6     7     7     8 ...
+       9    10    11    12    13    12    13    14    15    16 ...
+      17    17    18    19    20    21    22    23    23    24 ...
+      25    26    27    27    28    29    30    31    31    32 ...
+      33    34    35    35    36    37    38    39    40    40 ...
+      41    42    43    44    44    45    46    47    48    49 ...
+      49    50    51    52    53    54    55    56    57    58 ...
+      59    59    60    61    62    63    63    64    65    66 ...
+      67    67    68    69    70    71    72    72    73    74 ...
+      75    76    76    77    78    79    80    81    81    82 ...
+      83    84    85    85    86    87    88    89    90    90 ...
+      91    92    93    94    94    95    96    97    98    98 ...
+      99   100   101   102   103   103   104   106   107   108 ...
+     108   109   110   111   112   113   113   114   115   116 ...
+     117   117   118   119   120   121   122   122   123   124 ...
+     125   126   126   127   128   129   130   130   131   132 ...
+     133   134   135   135   136   137   138   139   139   140 ...
+     141   142   143   144   144   145   146   147   148   148 ...
+     149   150   151   152   153   153   154   155   156   158 ...
+     158   159   160   161   162   162   163   164   165   166 ...
+    ]/2.25;
+jd = [ ...
+       3     6    10    14    18    22    25    29    32    36 ...
+      40    43    47    50    54    58    61    65    68    72 ...
+      76    80    84    87    91    95    98   102   105   109 ...
+     112   116   120   124   128   131   135   138   142   146 ...
+     149   153   157   160   164   167   171   175   178   182 ...
+     186   189   193   197   200   204   207   211   215   218 ...
+     222   226   229   233   236   241   245   248   252   256 ...
+     259   263   267   270   274   277   281   285   288   292 ...
+     296   299   303   306   310   314   317   321   325   328 ...
+     332   336   339   343   346   350   354   357   361   365 ...
+     368   372   375   379   383   386   390   394   397   401 ...
+     405   408   412   415   419   423   426   430   434   437 ...
+     441   444   448   452   455   459   463   467   471   475 ...
+     478   482   485   489   493   496   500   504   507   511 ...
+     514   518   522   525   529   533   536   540   544   547 ...
+     551   554   558   562   565   569   573   576   580   583 ...
+     587   591   594   598   602   605   609   613   616   620 ...
+     623   627   631   634   638   642   645   649   652   656 ...
+     660   663   667   671   674   678   682   685   689   693 ...
+     697   701   704   708   712   715   719   722   726   730 ...
+    ]/2.25;
+
+% JavaScript font metrics
+wa = [ ...
+       1     4     2     4     5     7     8     9    11    11 ...
+      10    12    15    14    15    16    17    18    19    19 ...
+      22    24    25    23    25    28    27    29    30    32 ...
+      32    33    33    36    37    37    38    39    40    41 ...
+      42    43    44    46    46    46    47    50    50    51 ...
+      51    53    54    56    56    57    59    58    59    60 ...
+      63    63    64    65    66    67    68    70    71    73 ...
+      71    72    74    75    76    77    78    80    79    82 ...
+      84    82    85    85    86    88    88    89    91    92 ...
+      94    93    93    97    96    99    99    99   101   102 ...
+     101   105   106   106   106   107   111   110   110   112 ...
+     113   114   116   117   118   119   120   120   121   121 ...
+     123   124   126   127   128   129   130   129   133   134 ...
+     133   134   135   137   138   139   140   141   141   144 ...
+     143   146   145   146   148   149   151   151   152   154 ...
+     155   154   157   156   158   159   160   162   162   161 ...
+     165   166   166   168   167   169   170   171   173   173 ...
+     175   176   175   177   179   179   182   181   183   183 ...
+     184   186   187   189   190   188   190   193   192   194 ...
+     194   196   197   198   200   201   201   203   202   204 ...
+     ]/2.25;
+wd = [ ...
+       4     9    11    15    19    24    28    32    35    40 ...
+      41    45    51    54    58    62    66    70    73    76 ...
+      82    86    90    91    96   102   103   108   111   116 ...
+     120   123   126   131   135   138   141   146   149   153 ...
+     158   162   166   170   174   176   180   186   188   192 ...
+     195   200   204   208   212   216   220   222   226   230 ...
+     236   239   242   246   250   254   258   262   266   271 ...
+     272   276   280   284   288   292   296   301   302   308 ...
+     312   314   320   322   326   331   334   338   342   346 ...
+     351   352   356   363   364   370   372   376   381   384 ...
+     386   393   396   400   402   406   413   415   418   423 ...
+     427   430   435   438   443   447   450   452   456   460 ...
+     464   468   472   476   480   484   488   490   496   500 ...
+     503   506   510   514   518   522   526   530   533   538 ...
+     540   546   548   552   556   560   565   568   572   576 ...
+     580   582   588   590   595   598   602   606   610   612 ...
+     618   622   625   630   632   636   640   644   648   652 ...
+     657   660   662   666   672   675   680   682   687   690 ...
+     694   698   702   707   710   712   717   722   724   728 ...
+     732   737   740   744   749   752   756   761   762   767 ...
+     ]/2.25;
+
+f = ancestor( obj, 'figure' );
+s = min( obj.FontSize, 200 );
+if isempty( f ) || ~isprop( f, 'JavaFrame_I' ) || ~isempty( f.JavaFrame_I )
+    a = ja(s);
+    d = jd(s);
+else
+    a = wa(s);
+    d = wd(s);
+end
+
+end % fontmetrics